requirements.txt
fastapi
uvicorn
google-generativeai
tweepy
python-dotenv
pyyaml
requests
psycopg2-binary
openai
unidecode